+++ src/calliope/backend/gurobi_backend_model.py
@@ -149,7 +149,7 @@ class GurobiBackendModel(backend_model.BackendModel):
 
     def set_objective(self, name: str) -> None:  # noqa: D102, override
         to_set = self.objectives[name]
-        sense = self.OBJECTIVE_SENSE_DICT[to_set.attrs["sense"]]
+        sense = self.OBJECTIVE_SENSE_DICT[self.math.objectives[name].sense]
         self._instance.setObjective(to_set.item(), sense=sense)
         self._instance.update()
         self.objective = name