+++ src/calliope/preprocess/model_data.py
@@ -126,7 +126,6 @@ class ModelDataFactory:
         self.add_colors()
         self.add_link_distances()
         self.update_and_resample_dimensions()
-        self.assign_input_attr()
         self.dataset = self.dataset.assign_coords(
             self._update_dtypes(self.dataset.coords)
         )
@@ -383,17 +382,6 @@ class ModelDataFactory:
                 "Filling missing technology color array values from default palette."
             )
             self.dataset["color"] = self.dataset["color"].fillna(new_color_array)
-
-    def assign_input_attr(self):
-        """Assign metadata as attributes to each input array."""
-        all_attrs = {
-            **self.math.parameters.model_dump(),
-            **self.math.lookups.model_dump(),
-        }
-        for var_name, var_data in self.dataset.data_vars.items():
-            self.dataset[var_name] = var_data.assign_attrs(all_attrs.get(var_name, {}))
-            # Remove this redundant attribute
-            self.dataset[var_name].attrs.pop("active", None)
 
     def _get_relevant_node_refs(self, techs_dict: AttrDict, node: str) -> list[str]:
         """Get all references to input data made in technologies at nodes.