+++ tests/test_preprocess_model_data.py
@@ -287,18 +287,6 @@ class TestModelData:
             ["#123", "#17344c", "#321", "#456"],
         )
 
-    def test_assign_input_attr(
-        self, model_data_factory: ModelDataFactory, simple_da, timeseries_da
-    ):
-        model_data_factory.dataset["storage_cap_max"] = simple_da
-        model_data_factory.dataset["bar"] = timeseries_da
-        assert model_data_factory.dataset.data_vars
-
-        model_data_factory.assign_input_attr()
-
-        assert model_data_factory.dataset["storage_cap_max"].attrs["default"] == np.inf
-        assert "default" not in model_data_factory.dataset["bar"].attrs
-
     def test_get_relevant_node_refs_ts_data(self, model_data_factory: ModelDataFactory):
         techs_dict = AttrDict(
             {